requirements.txt
python-socketio==5.13.0
python-dotenv==1.1.1
PyJWT==2.10.1
Flask==3.1.1